spark-consumer/requirements.txt
kafka-python==2.2.15
py4j==0.10.9.7
pyspark==3.5.1
python-dotenv==1.1.1
